LX Header Segmentation added

+++ JellyBins.LinearExecutable/Headers/LeExportEntry.cs
@@ -3,7 +3,7 @@
 public class LeExportEntry
 {
     public String? Name { get; set; }
-    public UInt32 Ordinal { get; set; }
+    public UInt16 Ordinal { get; set; }
     public UInt32 ObjectIndex { get; set; }
     public UInt32 Offset { get; set; }
 }
